requirements.txt
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.2
matplotlib==3.8.3
pillow==10.2.0